pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "alfbot"
version = "0.1.0"
description = "ALF bots"
requires-python = ">=3.8"

dependencies = ["coloralf", "python-telegram-bot"]

[tool.setuptools.packages.find]
where = ["."]